requirements.txt
lxml==4.9.1
requests==2.25.1
click==8.1.3
python-decouple==3.6